DifferentiableHarmonics.jl

This package is spun of QG3.jl to give a stand-alone implementation of reverse-mode differentiable spherical harmonics transforms and derivative operators that work on CUDA GPUs and CPU. All implementations are checked again finite difference implenations and the transforms of SpeedyWeather.jl.

How-to

Each of the transforms and derivatives needs to be pre-planned. The common parameters of these plans are hold in HarmonicsParameters. First, we can plan transforms for a quadratic FullGaussianGrid with truncation at the 21st wave number

L_max = 21
p = HarmonicsParameters(L_max, GPU=false)

Then, we generate some random fields in both the spectral and grid domain::

A = rand_grid(p)
B = rand_SH(p)

# batched matrices with trailing dimensions can be declared by:
A_b = rand_grid(p, N_channels, N_batch) # N_lat x N_lon x N_channels x N_batch
B_b = rand_SH(p,  N_channels, N_batch)

Now we can preplan all derivatives and transforms:

dλ = Derivative_dλ(p, N_batch=N_batch)
dμ = GaussianGrid_dμ(p, N_channels, N_batch)
L = Laplacian(p, init_inverse=true, N_batch=N_batch)
analysis_plan = SHtoGaussianGridTransform(p, N_channels, N_batch)
synthesis_plan = GaussianGridtoSHTransform(p, N_channels, N_batch)

And apply to plans like this:

transform_grid(B_b, analysis_plan) # SH -> grid 
transform_grid(A_b, synthesis_plan) # grid -> SH 
Δ(B_b, L) # SH -> Laplacian(SH)
SHtoGrid_dθ(B_b, dμ) # SH -> d B / dθ   colatitude
SHtoGrid_dμ(B_b, dμ) # SH -> d B / dμ   (μ = sin(lat))
SHtoGrid_dλ(B_b, dλ) # SH -> d B / dλ   longitutde

Harmonics Coefficients

The harmonics coefficents are saved in the following format:

show(sph_modes(p))

Work in Progress

In the future, this will be replaced by SpeedyWeather.jl's SpeedyTransform, but so far it is not differentiable yet.

DifferentiableHarmonics.AbstractSHTransform

DifferentiableHarmonics.GaussianGridtoSHTransform

DifferentiableHarmonics.GaussianGridtoSHTransform

DifferentiableHarmonics.HarmonicsParameters

DifferentiableHarmonics.HarmonicsParameters

DifferentiableHarmonics.HarmonicsParameters

DifferentiableHarmonics.SHtoGaussianGridTransform

DifferentiableHarmonics.SH_mask

DifferentiableHarmonics.compute_GaussWeights

DifferentiableHarmonics.compute_LegendreGauss

DifferentiableHarmonics.compute_LegendreGauss

DifferentiableHarmonics.compute_P

DifferentiableHarmonics.lMatrix

DifferentiableHarmonics.mMatrix

DifferentiableHarmonics.no_SH_mask

DifferentiableHarmonics.plan_ir2r_AD

DifferentiableHarmonics.plan_r2r_AD

DifferentiableHarmonics.rand_SH

DifferentiableHarmonics.rand_grid

DifferentiableHarmonics.sph_modes

DifferentiableHarmonics.transform_SH

DifferentiableHarmonics.transform_grid

DifferentiableHarmonics.zeros_SH

DifferentiableHarmonics.zeros_grid

DifferentiableHarmonics.AbstractSHTransform — Type

abstract type AbstractSHTransform{onGPU}

Required fields for all subtypes:

outputsize

DifferentiableHarmonics.GaussianGridtoSHTransform — Type

GaussianGridtoSHTransform(p::QG3ModelParameters{T}, N_level::Int=3; N_batch::Int=0)

DifferentiableHarmonics.GaussianGridtoSHTransform — Method

(

GaussianGridtoSHTransform(
    p::HarmonicsParameters{T}
) -> GaussianGridtoSHTransform{_A, _B, _C, _D, <:AbstractArray{T, 3}, Tuple{Int64, Int64}} where {_A, _B, _C, _D, T<:Number}
GaussianGridtoSHTransform(
    p::HarmonicsParameters{T},
    N_channels::Int64
) -> GaussianGridtoSHTransform{_A, _B, _C, _D, <:AbstractArray{T, 3}, Tuple{Int64, Int64}} where {_A, _B, _C, _D, T<:Number}
GaussianGridtoSHTransform(
    p::HarmonicsParameters{T},
    N_channels::Int64,
    N_batch::Int64
) -> GaussianGridtoSHTransform{_A, _B, _C, _D, <:AbstractArray{T, 3}, Tuple{Int64, Int64}} where {_A, _B, _C, _D, T<:Number}

)

Returns transform struct, that can be used with transform_SH. Transforms Gaussian Grid data to real spherical harmonics coefficients that follow the coefficient logic explained in the main documenation.

Additional input arguments:

N_channels: defines the transform for N_channel horizontal/channel levels. Has to be equal to three for the QG3 model itself, but might be different for other applications.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted

DifferentiableHarmonics.HarmonicsParameters — Type

HarmonicsParameters(
    Lmax::Integer;
    ...
) -> HarmonicsParameters{Float32, _A, <:Tuple{Any, Any}} where _A
HarmonicsParameters(
    Lmax::Integer,
    antialiasing::Symbol;
    eltype,
    GPU
) -> HarmonicsParameters{Float32, _A, <:Tuple{Any, Any}} where _A

Initializes the HarmonicsParameters based on a standard triangular truncation, :quadratic or :cubic antialiasing and Gaussian latitudes.

DifferentiableHarmonics.HarmonicsParameters — Type

Holds all parameters for the spherical harmonics. L and M are the number of harmonics numbers. L_max is the largest value of l.

DifferentiableHarmonics.HarmonicsParameters — Method

HarmonicsParameters(
    Lmax::Integer,
    lats::AbstractArray{T, 1},
    lons::AbstractArray{T, 1};
    GPU
)

Initializes the HarmonicsParameters based on a standard triangular truncation and the grid handed over.

DifferentiableHarmonics.SHtoGaussianGridTransform — Method

SHtoGaussianGridTransform(
    p::HarmonicsParameters{T}
) -> SHtoGaussianGridTransform{_A, _B, _C, _D, _E, Tuple{Int64, Int64}, Int64} where {_A, _B, _C, _D, _E}
SHtoGaussianGridTransform(
    p::HarmonicsParameters{T},
    N_channels::Int64
) -> SHtoGaussianGridTransform{_A, _B, _C, _D, _E, Tuple{Int64, Int64}, Int64} where {_A, _B, _C, _D, _E}
SHtoGaussianGridTransform(
    p::HarmonicsParameters{T},
    N_channels::Int64,
    N_batch::Int64
) -> SHtoGaussianGridTransform{_A, _B, _C, _D, _E, Tuple{Int64, Int64}, Int64} where {_A, _B, _C, _D, _E}

Returns transform struct, that can be used with transform_grid. Transforms real spherical harmonics coefficients to Gaussian grid data, follows the coefficient logic explained in the main documenation.

N_channel: defines the transform for N_channel horizontal levels.

N_batch: defines the transforms with an additional batch dimension for ML tasks, if N_batch==0 this is omitted and only 2D and 3D input will work

DifferentiableHarmonics.SH_mask — Method

(

SH_mask(p::HarmonicsParameters) -> Any

) Returns a Boolean array with all entries that represent SPH coefficents true

DifferentiableHarmonics.compute_GaussWeights — Method

(

compute_GaussWeights(
    p::HarmonicsParameters{T<:Number}
) -> Any
compute_GaussWeights(
    p::HarmonicsParameters{T<:Number},
    reltol
) -> Any

) Pre-computes gaussian weights for Legendre Transform, also checks if we really have the correct gaussian latitudes

DifferentiableHarmonics.compute_LegendreGauss — Method

(

compute_LegendreGauss(
    p::HarmonicsParameters{T<:Number},
    P::AbstractArray{T<:Number, 3},
    w::AbstractArray{T<:Number, 1}
) -> AbstractArray{T, 3} where T<:Number

)

DifferentiableHarmonics.compute_LegendreGauss — Method

(

compute_LegendreGauss(
    p::HarmonicsParameters{T<:Number},
    P::AbstractArray{T<:Number, 3};
    reltol
) -> AbstractArray{T, 3} where T<:Number

)

DifferentiableHarmonics.compute_P — Method

compute_P(
    p::HarmonicsParameters{T<:Number};
    sh_norm,
    CSPhase,
    prefactor
) -> Tuple{Any, Any}

Pre-compute ass. Legendre Polynomials and dP/dx (derivative of ass. Legendre Polynomial) at the grid points and also the remainder of the Spherical Harmonics at the grid points using GSL

m values are stored 0,1,2,3,4,5,6,7, ...l_max, 0 (nothing),-1, -2, -3, (on GPU) (the second 0 is the Imanigary part / sin part of the fourier transform which is always identical to zero, it is kept here to have equal matrix sizes)

so far only |m| is used, as I assume real SPH.

DifferentiableHarmonics.lMatrix — Method

(

lMatrix(p::HarmonicsParameters) -> Any

)

Returns the order numbers l of the SPH coefficents.

DifferentiableHarmonics.mMatrix — Method

(

mMatrix(p::HarmonicsParameters) -> Any

)

Returns the order numbers m of the SPH coefficents.

DifferentiableHarmonics.no_SH_mask — Method

(

no_SH_mask(p::HarmonicsParameters) -> Any

) Returns a Boolean array with all entries that don't represent SPH coefficents true

DifferentiableHarmonics.plan_ir2r_AD — Function

plan_ir2r_AD(arr::AbstractArray{T,N}, dims=1)

Returns an inverse R2R half-complex format FFT plan (HC2R) along dimension dims. Incontrast to the FFTW.R2R plan this assumes both the real and imaginary part in ascending order, so (Re(f0), Re(f1), .... , Im(f0), Im(f1), ...

Similar to the R2R plan, this plan is not normalized. For a normaliziation you have to divide by the number of elements in the dimension you apply the plan to.

DifferentiableHarmonics.plan_r2r_AD — Function

plan_r2r_AD(arr::AbstractArray{T,N}, dims=1)

Returns a R2R half-complex format FFT plan (R2HC) along dimension dims. Incontrast to the FFTW.R2R plan this returns both the real and imaganiary part in ascending order, so (Re(f0), Re(f1), .... , Im(f0), Im(f1), ...

DifferentiableHarmonics.rand_SH — Method

(

rand_SH(p::HarmonicsParameters{T}, size...) -> Any

)

Generate a random array with the correct shape to represend the SH domain and on the correct device and eltype. Additional input arguments are appended after the first two leading dimensions

DifferentiableHarmonics.rand_grid — Method

(

rand_grid(p::HarmonicsParameters{T}, size...) -> Any

)

Generate a zero array with the correct shape to represend the grid domain and on the correct device and eltype. Additional input arguments are appended after the first two leading dimensions

DifferentiableHarmonics.sph_modes — Method

(

sph_modes(p::HarmonicsParameters) -> Any

)

Returns a matrix with the SPH modes (l, m) arranged as they are in all SH domain matrices in DifferentiableHarmonics.jl

DifferentiableHarmonics.transform_SH — Function

transform_SH(data::AbstractArray{T,N}, t::GaussianGridtoSHTransform)

Transforms data into the spherical harmonics domain. The coefficents are ordered in a matrix in coloumns of the m value. On CPU the convention of FastTransform.jl is used (0, -1, 1, -2, 2, ...), on GPU the convention (0, 1, 2, 3, ...., (nothing, -1, -2, -3, ...)). Watch out, in future proabaly this might be standardized.

DifferentiableHarmonics.transform_grid — Function

transform_grid(data::AbstractArray{T,N}, t::SHtoGaussianGridTransform)

Transforms data from the spherical harmonics domain to a Gaussian Grid. The coefficents are ordered in a matrix in coloumns of the m value. On CPU the convention of FastTransform.jl is used (0, -1, 1, -2, 2, ...), on GPU the convention (0, 1, 2, 3, ...., (nothing, -1, -2, -3, ...)). Watch out, in future proabaly this might be standardized.

DifferentiableHarmonics.zeros_SH — Method

(

zeros_SH(p::HarmonicsParameters{T}, size...) -> Any

)

Generate a zero array with the correct shape to represend the SH domain and on the correct device and eltype. Additional input arguments are appended after the first two leading dimensions

DifferentiableHarmonics.zeros_grid — Method

(

zeros_grid(p::HarmonicsParameters{T}, size...) -> Any

)

Generate a zero array with the correct shape to represend the grid domain and on the correct device and eltype. Additional input arguments are appended after the first two leading dimensions